requirements.txt
pydantic>=2.0.0
pyyaml>=6.0
requests>=2.31.0
python-dotenv>=1.0.0
feedparser>=6.0.10
tweepy>=4.14.0
google-genai>=0.3.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
